+++ backend/src/main/java/skkuchin/service/service/ChatRoomService.java
@@ -10,8 +10,6 @@ import skkuchin.service.domain.Chat.ChatMessage;
 import skkuchin.service.domain.Chat.ChatRoom;
 import skkuchin.service.domain.Chat.ResponseType;
 import skkuchin.service.domain.User.AppUser;
-import skkuchin.service.dto.DebeziumDto;
-import skkuchin.service.dto.UserDto;
 import skkuchin.service.exception.CustomRuntimeException;
 import skkuchin.service.repo.ChatMessageRepo;
 import skkuchin.service.repo.ChatRoomRepo;
@@ -20,6 +18,7 @@ import skkuchin.service.repo.UserRepo;
 import javax.transaction.Transactional;
 import java.time.LocalDateTime;
 import java.util.*;
+import java.util.stream.Collectors;
 
 @Service
 @Slf4j
@@ -29,10 +28,8 @@ public class ChatRoomService {
     private final ChatRoomRepo chatRoomRepo;
     private final ChatMessageRepo chatMessageRepo;
     private final UserRepo userRepo;
-    private final ChatMessageService chatMessageService;
-    private final UserService userService;
 
-    //채팅방의 메시지들 조회
+    @Transactional
     public void updateReadStatus(ChatRoom chatRoom, String sender){
         List<ChatMessage> chatMessages = chatMessageRepo.findByChatRoom(chatRoom);
         for (ChatMessage chatMessage : chatMessages) {
@@ -63,6 +60,61 @@ public class ChatRoomService {
 
         chatRoom.setResponse(responseType);
         chatRoomRepo.save(chatRoom);
+    }
+
+    @Transactional
+    public List<ChatRoomDto.Response> getChatRoomList(String username){
+
+        AppUser user = userRepo.findByUsername(username);
+        List<ChatRoom> chatRooms = chatRoomRepo.findMyRoomList(user.getId());
+
+        if (chatRooms.size() > 0) {
+            List<ChatRoomDto.Response> chatRoomDtos = chatRooms
+                    .stream()
+                    .map(chatRoom -> new ChatRoomDto.Response(
+                            chatRoom,
+                            getLatestMessage(chatRoom),
+                            unReadMessage(chatRoom.getRoomId(), username),
+                            getOtherUser(chatRoom, username)
+                    ))
+                    .collect(Collectors.toList());
+            Collections.sort(chatRoomDtos,new DateComparator().reversed());
+
+            return chatRoomDtos;
+        }
+        List<ChatRoomDto.Response> chatRoomDto = new ArrayList<>();
+        return chatRoomDto;
+    }
+
+    @Transactional
+    public List<ChatRoomDto.userResponse> getAlarmList(String username){
+        AppUser user = userRepo.findByUsername(username);
+        List<ChatRoom> chatRooms = chatRoomRepo.findRequestByUserId(user.getId());
+        List<ChatRoomDto.userResponse> chatRoomDto = chatRooms
+                .stream()
+                .map(chatRoom -> new ChatRoomDto.userResponse(chatRoom))
+                .collect(Collectors.toList());
+        Collections.sort(chatRoomDto, new AlarmListDateComparator().reversed());
+        return chatRoomDto;
+    }
+
+    @Transactional
+    public AppUser getOtherUser(ChatRoom chatRoom, String username){
+        AppUser user = userRepo.findByUsername(username);
+        return chatRoomRepo.findOtherUser(chatRoom, user.getId());
+    }
+
+    private int unReadMessage(String roomId, String username){
+        return chatMessageRepo.countByReadStatus(roomId, username);
+    }
+
+    private ChatMessage getLatestMessage(ChatRoom chatRoom){
+        List<ChatMessage> chatMessages = chatMessageRepo.findByLatestTime(chatRoom.getRoomId());
+        if(chatMessages.size() == 0) {
+            throw new CustomRuntimeException("새로운 채팅방이 개설되었습니다");
+        }
+        ChatMessage chatMessage = chatMessages.get(0);
+        return chatMessage;
     }
 
 
@@ -139,20 +191,32 @@ public class ChatRoomService {
         }
     }
 
+    @Transactional
     public ChatRoomDto.settingResponse getSettingResponse(ChatRoom chatRoom){
         return new ChatRoomDto.settingResponse(chatRoom);
     }
 
-    public ChatRoomDto.userResponse getUserRoomDto(ChatRoom chatRoom){
-        return new ChatRoomDto.userResponse(chatRoom);
+    private class DateComparator implements Comparator<ChatRoomDto.Response> {
+        @Override
+        public int compare(ChatRoomDto.Response f1, ChatRoomDto.Response f2) {
+            if (f1.getMessageTime().isAfter(f2.getMessageTime()) ) {
+                return 1;
+            } else  {
+                return -1;
+            }
+        }
     }
 
-//    public DebeziumDto.UserChatInfo getUserChatInfo(String username){
-//        UserDto.chatRoomResponse userInfo = userService.getChatRoomUser(username);
-//        List<ChatRoomDto.Response> userChatMessages = chatMessageService.getChatList(username);
-//        DebeziumDto.UserChatInfo userChatInfo = new DebeziumDto.UserChatInfo(userInfo,userChatMessages);
-//        return userChatInfo;
-//    }
+    private class AlarmListDateComparator implements Comparator<ChatRoomDto.userResponse> {
+        @Override
+        public int compare(ChatRoomDto.userResponse f1, ChatRoomDto.userResponse f2) {
+            if (f1.getCreatedDate().isAfter(f2.getCreatedDate()) ) {
+                return 1;
+            } else  {
+                return -1;
+            }
+        }
+    }
 
 }
