+++ backend/src/main/java/skkuchin/service/webSocket/controller/DebeziumController.java
@@ -15,10 +15,6 @@
 //import skkuchin.service.repo.ChatRoomRepo;
 //import skkuchin.service.service.ChatMessageService;
 //import skkuchin.service.service.ChatRoomService;
-//import skkuchin.service.service.UserService;
-//
-//import java.time.Duration;
-//import java.time.LocalDateTime;
 //import java.util.List;
 //
 //@RestController
@@ -29,20 +25,16 @@
 //    private final ChatRoomService chatRoomService;
 //    private final ChatMessageService chatMessageService;
 //    private final ChatRoomRepo chatRoomRepo;
-//    private final UserService userService;
-//
-//
 //
 //    @KafkaListener(topics = "dbserver.service.chat_message")
 //    public void listenChatMessage(@Payload(required = false) String message) throws Exception {
 //        System.out.println("kafka consume test topic : "  + message);
-//        LocalDateTime start = LocalDateTime.now();
-//        //payload null 여부 확인
+//
 //        ObjectMapper objectMapper = new ObjectMapper();
 //        objectMapper.configure(DeserializationFeature.FAIL_ON_UNKNOWN_PROPERTIES, false);
 //        DebeziumDto.ChatMessageRequest dto= objectMapper.readValue(message, DebeziumDto.ChatMessageRequest.class);
-//
 //        ChatRoom chatRoom = chatRoomService.findChatById(dto.getPayload().getAfter().getChat_room_id());
+//
 //        AppUser user1 = chatRoom.getUser1();
 //        AppUser user2 = chatRoom.getUser2();
 //
@@ -52,53 +44,28 @@
 //        String user2Name = user2.getUsername();
 //
 //        List<ChatMessageDto.Response> chatMessages = chatMessageService.getAllMessage(chatRoom);
-//        for (int i = 0; i <chatMessages.size() ; i++) {
-//            System.out.println("------index------- : " + i);
-//            System.out.println("chatMessages.message = " + chatMessages.get(i).getMessage());
-//            System.out.println("chatMessages.date = " + chatMessages.get(i).getDate());
-//            System.out.println("chatMessages.sender = " + chatMessages.get(i).getSender());
-//            System.out.println("chatMessages.roomId = " + chatMessages.get(i).getRoomId());
-//        }
 //
 //        template.convertAndSend(CHAT_EXCHANGE_NAME, "chat." + roomId+"user1", chatMessages);
 //        template.convertAndSend(CHAT_EXCHANGE_NAME, "chat." + roomId+"user2", chatMessages);
 //
-//        List<ChatRoomDto.Response> chatRooms1 = chatMessageService.getChatList(user1Name);
-//        List<ChatRoomDto.Response> chatRooms2 = chatMessageService.getChatList(user2Name);
-//
-//        for (int i = 0; i < chatRooms1.size(); i++) {
-//            System.out.println("------index------- : " + i);
-//            System.out.println("chatRooms1.messageCount = " + chatRooms1.get(i).getMessageCount());
-//            System.out.println("chatRooms1.messageTime = " + chatRooms1.get(i).getMessageTime());
-//            System.out.println("chatRooms1.roomId = " + chatRooms1.get(i).getRoomId());
-//            System.out.println("chatRooms1.message = " + chatRooms1.get(i).getMessage());
-//        }
-//
-//        for (int i = 0; i < chatRooms2.size(); i++) {
-//            System.out.println("------index------- : " + i);
-//            System.out.println("chatRooms2.messageCount = " + chatRooms2.get(i).getMessageCount());
-//            System.out.println("chatRooms2.messageTime = " + chatRooms2.get(i).getMessageTime());
-//            System.out.println("chatRooms2.roomId = " + chatRooms2.get(i).getRoomId());
-//            System.out.println("chatRooms2.message = " + chatRooms2.get(i).getMessage());
-//        }
-//
+//        List<ChatRoomDto.Response> chatRooms1 = chatRoomService.getChatRoomList(user1Name);
+//        List<ChatRoomDto.Response> chatRooms2 = chatRoomService.getChatRoomList(user2Name);
 //
 //        template.convertAndSend(CHAT_EXCHANGE_NAME,"room."+user1Name+"chatRoomList", chatRooms1);
 //        template.convertAndSend(CHAT_EXCHANGE_NAME,"room."+user2Name+"chatRoomList", chatRooms2);
-//        System.out.println("Duration.between(LocalDateTime.now(),start) = " + Duration.between(LocalDateTime.now(),start));
 //    }
 //
 //
 //    @KafkaListener(topics = "dbserver.service.chat_room")
 //    public void chatRoomChange(@Payload(required = false) String message) throws Exception {
 //        System.out.println("kafka consume test topic : "  + message);
+//
 //        ObjectMapper objectMapper = new ObjectMapper();
 //        objectMapper.configure(DeserializationFeature.FAIL_ON_UNKNOWN_PROPERTIES, false);
 //        DebeziumDto.ChatRoomRequest dto= objectMapper.readValue(message, DebeziumDto.ChatRoomRequest.class);
 //
 //        String roomId = dto.getPayload().getAfter().getRoom_id();
 //        ChatRoom chatRoom = chatRoomService.findChatRoom(roomId);
-//        ChatRoomDto.settingResponse chatRoomBlockInfo = chatRoomService.getSettingResponse(chatRoom);
 //
 //        AppUser user1 = chatRoom.getUser1();
 //        AppUser user2 = chatRoom.getUser2();
@@ -106,65 +73,52 @@
 //        String userName1 = user1.getUsername();
 //        String userName2 = user2.getUsername();
 //
-//        List<ChatRoomDto.Response> chatRooms1 = chatMessageService.getChatList(userName1);
-//        List<ChatRoomDto.Response> chatRooms2 = chatMessageService.getChatList(userName2);
-//        List<ChatRoomDto.userResponse> alarmList1 = chatMessageService.getAlarmList(userName1);
-//        List<ChatRoomDto.userResponse> alarmList2 = chatMessageService.getAlarmList(userName2);
+//        List<ChatRoomDto.Response> chatRooms1 = chatRoomService.getChatRoomList(userName1);
+//        List<ChatRoomDto.Response> chatRooms2 = chatRoomService.getChatRoomList(userName2);
+//        ChatRoomDto.settingResponse settingResponse = chatRoomService.getSettingResponse(chatRoom);
 //
 //        if(dto.getPayload().getOp().equals("c")){
-//            String username = user2.getUsername();
-//            List<ChatRoomDto.userResponse> alarmList = chatMessageService.getAlarmList(username);
-//            template.convertAndSend(CHAT_EXCHANGE_NAME,"alarm."+ username,alarmList);
+//            List<ChatRoomDto.userResponse> alarmList = chatRoomService.getAlarmList(userName2);
+//            template.convertAndSend(CHAT_EXCHANGE_NAME,"alarm."+ userName2, alarmList);
 //        }
 //        template.convertAndSend(CHAT_EXCHANGE_NAME, "room."+userName1+"chatRoomList", chatRooms1);
 //        template.convertAndSend(CHAT_EXCHANGE_NAME, "room."+userName2+"chatRoomList", chatRooms2);
-//        template.convertAndSend(CHAT_EXCHANGE_NAME,"setting."+chatRoomId +"user2",settingResponse);
-//        template.convertAndSend(CHAT_EXCHANGE_NAME,"setting."+chatRoomId +"user2",settingResponse);
+//        template.convertAndSend(CHAT_EXCHANGE_NAME,"setting."+roomId +"user1",settingResponse);
+//        template.convertAndSend(CHAT_EXCHANGE_NAME,"setting."+roomId +"user2",settingResponse);
 //    }
 //
 //
 //    @KafkaListener(topics = "dbserver.service.user")
 //    public void userChange(@Payload(required = false) String message) throws Exception{
 //        System.out.println("kafka consume test topic : "  + message);
-//        //payload null 여부 확인
 //
 //        ObjectMapper objectMapper = new ObjectMapper();
 //        objectMapper.configure(DeserializationFeature.FAIL_ON_UNKNOWN_PROPERTIES, false);
 //        DebeziumDto.UserRequest dto= objectMapper.readValue(message, DebeziumDto.UserRequest.class);
-///*
-//        String userName1 = dto.getPayload().getAfter().getUsername();
-//*/
+//
 //        Long userId = dto.getPayload().getAfter().getId();
-//        UserDto.Response userDto = userService.getUser(userId);
 //
-//        System.out.println("userDto.getUsername() = " + userDto.getUsername());
-//        System.out.println("userDto.getMajor() = " + userDto.getMajor());
-//        if(dto.getPayload().getOp().equals("u") || dto.getPayload().getOp().equals("d")){
-//
-//            /*template.convertAndSend(CHAT_EXCHANGE_NAME,
-//                    "alarm."+userName1,dto);
-//*/
+//        if (dto.getPayload().getOp().equals("u") || dto.getPayload().getOp().equals("d")) {
 //            List<ChatRoom> chatRooms = chatRoomRepo.findMyRoomList(userId);
-//            for (int i = 0; i < chatRooms.size(); i++) {
-//                AppUser user1 = chatRoomService.findUser1(chatRooms.get(i));
-//                AppUser user2 = chatRoomService.findUser2(chatRooms.get(i));
+//
+//            for (ChatRoom chatRoom : chatRooms) {
+//                AppUser user1 = chatRoom.getUser1();
+//                AppUser user2 = chatRoom.getUser2();
+//
 //                String user1Name = user1.getUsername();
 //                String user2Name = user2.getUsername();
-//                DebeziumDto.UserChatInfo user1ChatInfo = chatRoomService.getUserChatInfo(user1Name);
-//                DebeziumDto.UserChatInfo user2ChatInfo = chatRoomService.getUserChatInfo(user2Name);
 //
-//                System.out.println("user2ChatInfo.getUserInfo().getUsername() = " + user2ChatInfo.getUserInfo().getUsername());
-//                System.out.println("user2ChatInfo.getUserInfo().getMajor() = " + user2ChatInfo.getUserInfo().getMajor());
-//                template.convertAndSend(CHAT_EXCHANGE_NAME,
-//                            "room."+user1Name+"chatRoomList",user1ChatInfo);
-//                template.convertAndSend(CHAT_EXCHANGE_NAME,
-//                        "room."+user1Name+"chatRoomList",user2ChatInfo);
+//                List<ChatRoomDto.Response> chatRooms1 = chatRoomService.getChatRoomList(user1Name);
+//                List<ChatRoomDto.Response> chatRooms2 = chatRoomService.getChatRoomList(user2Name);
 //
+//                UserDto.Response user1Dto= new UserDto.Response(user1);
+//                UserDto.Response user2Dto = new UserDto.Response(user2);
 //
-//                template.convertAndSend(CHAT_EXCHANGE_NAME,
-//                        "user."+chatRooms.get(i).getRoomId()+"user1",userDto);
-//                template.convertAndSend(CHAT_EXCHANGE_NAME,
-//                        "user."+chatRooms.get(i).getRoomId()+"user2",userDto);
+//                template.convertAndSend(CHAT_EXCHANGE_NAME, "room."+user1Name+"chatRoomList", chatRooms1);
+//                template.convertAndSend(CHAT_EXCHANGE_NAME, "room."+user2Name+"chatRoomList", chatRooms2);
+//
+//                template.convertAndSend(CHAT_EXCHANGE_NAME, "user."+chatRoom.getRoomId()+"user1",user2Dto);
+//                template.convertAndSend(CHAT_EXCHANGE_NAME, "user."+chatRoom.getRoomId()+"user2",user1Dto);
 //            }
 //        }
 //    }
